Added placeholders for query assembly
