--- src/sql_orm/select/sort.js
@@ -1,1 +1,0 @@
-//handles sort operations
