+++ src/sql_orm/select/where/join_condition.js
@@ -1,0 +1,5 @@
+//how to join
+
+//array of arrays
+//objects of objects//arrays?
+    //join_type, arrays
